pyproject.toml
[tool.poetry]
name = "self-study-project"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.2.5"
djangorestframework = "^3.16.1"
djangorestframework-simplejwt = "^5.5.1"
drf-spectacular = "^0.28.0"
psycopg2-binary = "^2.9.10"
django-cors-headers = "^4.7.0"
python-dotenv = "^1.1.1"
pytest = "^8.4.1"
pytest-django = "^4.11.1"
django-filter = "^25.1"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.3.0"
black = "^25.1.0"
isort = "^6.0.1"
pre-commit = "^4.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 119
exclude = '''
(
  /(
      \.eggs         # Исключить несколько общих каталогов
    | \.git          # в корне проекта
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | dist
  )/
  | foo.py           # Также отдельно исключить файл с именем foo.py
                     # в корне проекта
)
'''

[tool.isort]
profile = "black"
line_length = 119
